+++ src/common/sql.c
@@ -1036,7 +1036,7 @@ void Sql_HerculesUpdateCheck(Sql* self) {
 		fseek (ufp,1,SEEK_SET);/* woo. skip the # */
 
 		if( fgets(timestamp,sizeof(timestamp),ufp) ) {
-			unsigned int timestampui = atol(timestamp);
+			unsigned int timestampui = (unsigned int)atol(timestamp);
 			if( SQL_ERROR == SQL->Query(self, "SELECT 1 FROM `sql_updates` WHERE `timestamp` = '%u' LIMIT 1", timestampui) )
 				Sql_ShowDebug(self);
 			if( Sql_NumRows(self) != 1 ) {
@@ -1079,7 +1079,7 @@ void Sql_HerculesUpdateSkip(Sql* self,const char *filename) {
 	fseek (ifp,1,SEEK_SET);/* woo. skip the # */
 	
 	if( fgets(timestamp,sizeof(timestamp),ifp) ) {
-		unsigned int timestampui = atol(timestamp);
+		unsigned int timestampui = (unsigned int)atol(timestamp);
 		if( SQL_ERROR == SQL->Query(self, "SELECT 1 FROM `sql_updates` WHERE `timestamp` = '%u' LIMIT 1", timestampui) )
 			Sql_ShowDebug(self);
 		else if( Sql_NumRows(self) == 1 ) {
